+++ branches/version_2_double/DrawEngine.Renderer/Mesh/MeshModel.cs
@@ -75,23 +75,23 @@ namespace DrawEngine.Renderer.Mesh
         #endregion
 
         #region ITransformable3D Members
-        public void Rotate(float angle, Vector3D axis)
+        public void Rotate(double angle, Vector3D axis)
         {
             throw new Exception("The method or operation is not implemented.");
         }
-        public void RotateAxisX(float angle)
+        public void RotateAxisX(double angle)
         {
             throw new Exception("The method or operation is not implemented.");
         }
-        public void RotateAxisY(float angle)
+        public void RotateAxisY(double angle)
         {
             throw new Exception("The method or operation is not implemented.");
         }
-        public void RotateAxisZ(float angle)
+        public void RotateAxisZ(double angle)
         {
             throw new Exception("The method or operation is not implemented.");
         }
-        public void Scale(float factor)
+        public void Scale(double factor)
         {
 
             if (this.accelerationManager == null)
@@ -108,7 +108,7 @@ namespace DrawEngine.Renderer.Mesh
             }
             
         }
-        public void Translate(float tx, float ty, float tz)
+        public void Translate(double tx, double ty, double tz)
         {   
             if (this.accelerationManager == null)
             {
@@ -198,7 +198,7 @@ namespace DrawEngine.Renderer.Mesh
         //        this.triangles = this.loader.Load();
         //        this.loader.Dispose();
         //        this.boundBox = this.loader.BoundBox;
-        //        float len = Math.Abs(this.boundBox.PMax.Y - this.boundBox.PMin.Y);
+        //        double len = Math.Abs(this.boundBox.PMax.Y - this.boundBox.PMin.Y);
         //        this.boundBox.Scale(50 / len);
         //        //this.boundBox.Translate(-this.boundBox.Center.ToVector3D());
         //        foreach (Triangle t in this.triangles)
